add: User accounts and Auth system

+++ routes/album.js
@@ -1,7 +1,22 @@
 const express = require("express");
 const { getSpotifyAccessToken } = require("./utils/spotify");
 const Album = require("../models/Albums");
+const { getAuth } = require("@clerk/express");
 const router = express.Router();
+
+function ensureAuthenticated(req, res, next) {
+    const { userId } = getAuth(req);
+
+    if (!userId) {
+        return res.status(401).json({ error: "Unauthorized" });
+    }
+
+    req.userId = userId;
+    next();
+}
+
+
+
 
 router.get("/album/:id", async(req, res) =>
 {
@@ -36,21 +51,27 @@ router.get("/album/:id", async(req, res) =>
     }
 });
 
-router.post("/album", async(req, res) => 
+router.post("/album", ensureAuthenticated, async(req, res) => 
 {
     try {
-        const album = await Album.create(req.body);
+        const album = await Album.create({ ...req.body, userId: req.userId });
         res.status(201).json(album);
     } catch (error) {
+        if (error.code === 11000) {
+            return res.status(409).json({ error: "Album already exists in your collection" });
+        }
+
         console.log(error);
         res.status(500).json({ error: "Failed to create album" });
     }
 });
 
-router.get("/album", async(req, res) =>
+router.get("/collection", ensureAuthenticated, async(req, res) =>
 {
+    const userId = req.userId;
+
     try {
-        const albums = await Album.find();
+        const albums = await Album.find({ userId });
         res.json(albums);
     } catch (error) {
         console.log(error);
@@ -58,9 +79,25 @@ router.get("/album", async(req, res) =>
     }
 });
 
-router.delete("/album/:id", async(req, res) => {
+router.get("/collections/:spotifyId", ensureAuthenticated, async(req, res) => {
     try {
-        await Album.findOneAndDelete({ spotifyId: req.params.id });
+        const  userId  = req.userId;
+        const { spotifyId } = req.params;
+
+        const album = await Album.findOne({ spotifyId, userId });
+        res.json({
+            saved: !!album,
+            album,
+        });
+    } catch (error) {
+            console.log(error);
+            res.status(500).json({ error: "Failed to check album in collection" });
+    }
+});
+
+router.delete("/album/:id", ensureAuthenticated, async(req, res) => {
+    try {
+        await Album.findOneAndDelete({ spotifyId: req.params.id, userId: req.userId });
         res.json({ message: "Album removed from collection" });
     } catch (error) {
         console.log(error)
